ripd, ripngd: Widen allow-ecmp leaf to uint16 for MULTIPATH_NUM paths
The allow-ecmp yang leaf and the rip/ripng "ecmp" field were uint8, but
MULTIPATH_NUM is 256. A bare "allow-ecmp" defaults to MULTIPATH_NUM, so
the northbound layer rejected the value.

Widen the leaf and the stored "ecmp" value to uint16 so the full
1..MULTIPATH_NUM range already advertised by the CLI range is representable.
Fixes the rip_allow_ecmp and ripng_allow_ecmp topotests.

+++ ripd/ripd.h
@@ -150,7 +150,7 @@ struct rip {
 	struct route_table *distance_table;
 
 	/* RIP ECMP flag */
-	uint8_t ecmp;
+	uint16_t ecmp;
 
 	/* Are we in passive-interface default mode? */
 	bool passive_default;
